refactor: urls.py, viewsets.py
Move urls.py and views.py, changed the api url
